[SOLR-15450] extract common methods of tests to base class
