+++ solr/contrib/ltr/src/test/org/apache/solr/ltr/TestLTROnSolrCloud.java
@@ -15,66 +15,17 @@
  */
 package org.apache.solr.ltr;
 
-import java.io.File;
-import java.util.Collections;
-import java.util.List;
-import java.util.SortedMap;
-import java.util.stream.IntStream;
-
-import org.apache.commons.io.FileUtils;
 import org.apache.solr.client.solrj.SolrQuery;
-import org.apache.solr.client.solrj.embedded.JettyConfig;
-import org.apache.solr.client.solrj.embedded.JettySolrRunner;
-import org.apache.solr.client.solrj.request.CollectionAdminRequest;
-import org.apache.solr.client.solrj.response.CollectionAdminResponse;
 import org.apache.solr.client.solrj.response.QueryResponse;
-import org.apache.solr.cloud.MiniSolrCloudCluster;
 import org.apache.solr.common.SolrInputDocument;
-import org.apache.solr.common.cloud.ZkStateReader;
 import org.apache.solr.ltr.feature.FieldValueFeature;
 import org.apache.solr.ltr.feature.OriginalScoreFeature;
 import org.apache.solr.ltr.feature.SolrFeature;
 import org.apache.solr.ltr.feature.ValueFeature;
 import org.apache.solr.ltr.model.LinearModel;
-import org.apache.solr.util.RestTestHarness;
-import org.eclipse.jetty.servlet.ServletHolder;
-import org.junit.AfterClass;
 import org.junit.Test;
 
-import static java.util.stream.Collectors.toList;
-
-public class TestLTROnSolrCloud extends TestRerankBase {
-
-  private MiniSolrCloudCluster solrCluster;
-  String solrconfig = "solrconfig-ltr.xml";
-  String schema = "schema.xml";
-
-  SortedMap<ServletHolder,String> extraServlets = null;
-
-  @Override
-  public void setUp() throws Exception {
-    super.setUp();
-    extraServlets = setupTestInit(solrconfig, schema, true);
-    System.setProperty("enable.update.log", "true");
-
-    int numberOfShards = random().nextInt(4)+1;
-    int numberOfReplicas = random().nextInt(2)+1;
-
-    int numberOfNodes = numberOfShards * numberOfReplicas;
-
-    setupSolrCluster(numberOfShards, numberOfReplicas, numberOfNodes);
-
-
-  }
-
-
-  @Override
-  public void tearDown() throws Exception {
-    restTestHarness.close();
-    restTestHarness = null;
-    solrCluster.shutdown();
-    super.tearDown();
-  }
+public class TestLTROnSolrCloud extends TestLTROnSolrCloudBase {
 
   @Test
   // commented 4-Sep-2018 @LuceneTestCase.BadApple(bugUrl="https://issues.apache.org/jira/browse/SOLR-12028") // 2-Aug-2018
@@ -218,43 +169,7 @@ public class TestLTROnSolrCloud extends TestRerankBase {
         queryResponse.getResults().get(7).get("features").toString());
   }
 
-  private void setupSolrCluster(int numShards, int numReplicas, int numServers) throws Exception {
-    JettyConfig jc = buildJettyConfig("/solr");
-    jc = JettyConfig.builder(jc).withServlets(extraServlets).build();
-    solrCluster = new MiniSolrCloudCluster(numServers, tmpSolrHome.toPath(), jc);
-    File configDir = tmpSolrHome.toPath().resolve("collection1/conf").toFile();
-    solrCluster.uploadConfigSet(configDir.toPath(), "conf1");
-
-    solrCluster.getSolrClient().setDefaultCollection(COLLECTION);
-
-    createCollection(COLLECTION, "conf1", numShards, numReplicas);
-    indexDocuments(COLLECTION);
-    for (JettySolrRunner solrRunner : solrCluster.getJettySolrRunners()) {
-      if (!solrRunner.getCoreContainer().getCores().isEmpty()){
-        String coreName = solrRunner.getCoreContainer().getCores().iterator().next().getName();
-        restTestHarness = new RestTestHarness(() -> solrRunner.getBaseUrl().toString() + "/" + coreName);
-        break;
-      }
-    }
-    loadModelsAndFeatures();
-  }
-
-
-  private void createCollection(String name, String config, int numShards, int numReplicas)
-      throws Exception {
-    CollectionAdminResponse response;
-    CollectionAdminRequest.Create create =
-        CollectionAdminRequest.createCollection(name, config, numShards, numReplicas);
-    response = create.process(solrCluster.getSolrClient());
-
-    if (response.getStatus() != 0 || response.getErrorMessages() != null) {
-      fail("Could not create collection. Response" + response.toString());
-    }
-    ZkStateReader zkStateReader = solrCluster.getSolrClient().getZkStateReader();
-    solrCluster.waitForActiveCollection(name, numShards, numShards * numReplicas);
-  }
-
-
+  @Override
   void indexDocument(String collection, String id, String title, String description, int popularity)
     throws Exception{
     SolrInputDocument doc = new SolrInputDocument();
@@ -274,27 +189,8 @@ public class TestLTROnSolrCloud extends TestRerankBase {
     solrCluster.getSolrClient().add(collection, doc);
   }
 
-  private void indexDocuments(final String collection)
-       throws Exception {
-    final int collectionSize = 8;
-    // put documents in random order to check that advanceExact is working correctly
-    List<Integer> docIds = IntStream.rangeClosed(1, collectionSize).boxed().collect(toList());
-    Collections.shuffle(docIds, random());
-
-    int docCounter = 1;
-    for (int docId : docIds) {
-      final int popularity = docId;
-      indexDocument(collection, String.valueOf(docId), "a1", "bloom", popularity);
-      // maybe commit in the middle in order to check that everything works fine for multi-segment case
-      if (docCounter == collectionSize / 2 && random().nextBoolean()) {
-        solrCluster.getSolrClient().commit(collection);
-      }
-      docCounter++;
-    }
-    solrCluster.getSolrClient().commit(collection, true, true);
-  }
-
-  private void loadModelsAndFeatures() throws Exception {
+  @Override
+  void loadModelsAndFeatures() throws Exception {
     final String featureStore = "test";
     final String[] featureNames = new String[]{"powpularityS", "c3", "original", "dvIntFieldFeature",
         "dvLongFieldFeature", "dvFloatFieldFeature", "dvDoubleFieldFeature", "dvStrNumFieldFeature", "dvStrBoolFieldFeature"};
@@ -366,21 +262,4 @@ public class TestLTROnSolrCloud extends TestRerankBase {
     );
     reloadCollection(COLLECTION);
   }
-
-  private void reloadCollection(String collection) throws Exception {
-    CollectionAdminRequest.Reload reloadRequest = CollectionAdminRequest.reloadCollection(collection);
-    CollectionAdminResponse response = reloadRequest.process(solrCluster.getSolrClient());
-    assertEquals(0, response.getStatus());
-    assertTrue(response.isSuccess());
-  }
-
-  @AfterClass
-  public static void after() throws Exception {
-    if (null != tmpSolrHome) {
-      FileUtils.deleteDirectory(tmpSolrHome);
-      tmpSolrHome = null;
-    }
-    System.clearProperty("managed.schema.mutable");
-  }
-
 }
